Back/API/requirements.txt
flask
requests
flask-cors
Werkzeug
flask_sqlalchemy
flask-migrate
psycopg2-binary
transformers
PyJWT==1.7.1
